refactor(page headline): move into its own namespace folder
