+++ src/modules/PageHeadline/PageHeadline.js
@@ -1,5 +1,5 @@
 import React from 'react'
-import styles from './PageHeadline/styles.module.css'
+import styles from './styles.module.css'
 
 export const PageHeadline = ({ children }) => (
     <h1 className={styles.headline}>{children}</h1>
